src/main/java/com/advancesd/group17/course/models/Course.java
package com.advancesd.group17.course.models;

import java.util.List;

import com.advancesd.group17.users.model.User;

public class Course {
	
	private int courseId;
	private String courseName;
	private String courseDesc;
	private Integer courseCredits;
	private User instructor;
	
	public Course() {
		super();
	}

	public Course(int courseId, String courseName) {
		super();
		this.courseId = courseId;
		this.courseName = courseName;
	}
	
	public Course(int courseId, String courseName, String courseDesc, Integer courseCredits, User instructor) {
		super();
		this.courseId = courseId;
		this.courseName = courseName;
		this.courseDesc = courseDesc;
		this.courseCredits = courseCredits;
		this.instructor = instructor;
		
	}

	public int getCourseId() {
		return courseId;
	}

	public void setCourseId(int courseId) {
		this.courseId = courseId;
	}

	public String getCourseName() {
		return courseName;
	}

	public void setCourseName(String courseName) {
		this.courseName = courseName;
	}

	public String getCourseDesc() {
		return courseDesc;
	}

	public void setCourseDesc(String courseDesc) {
		this.courseDesc = courseDesc;
	}

	public Integer getCourseCredits() {
		return courseCredits;
	}

	public void setCourseCredits(Integer courseCredits) {
		this.courseCredits = courseCredits;
	}

	public User getInstructor() {
		return instructor;
	}

	public void setInstructor(User instructor) {
		this.instructor = instructor;
	}


}
